criando a instacia de todas as camadas para a view e realizando teste para o consumo da api atraaves do debbug

+++ Movie/app/src/main/java/book/essencial/android/movie/api/Result.kt
@@ -17,6 +17,6 @@ data class Result(
     val release_date: String,
     val title: String,
     val video: Boolean,
-    val vote_average: Int,
+    val vote_average: Float,
     val vote_count: Int
 ) : Parcelable
